Initial commit for reorganization and restructuring of create garden form Removed 'multi-step' create garden form for now-- all on one page for easier development. Workshopping potential modal file upload flow
