+++ garden/src/components/form/garden/create/schemas.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-export const formSchema = z.object({
+export const gardenFormSchema = z.object({
   title: z
     .string()
     .min(1, { message: "Title is required" })
@@ -31,6 +31,55 @@ export const formSchema = z.object({
   publisher: z.string(),
   doi: z.string(),
   is_archived: z.boolean(),
+  modal: z
+    .object({
+      app_name: z.string(),
+      file_contents: z.string(),
+      modal_functions: z.array(
+        z.object({
+          title: z.string(),
+          description: z.string(),
+          function_name: z.string(),
+          is_archived: z.boolean(),
+          year: z.string(),
+          doi: z.string(),
+          function_text: z.string(),
+          authors: z.array(z.string()),
+          tags: z.array(z.string()),
+          test_functions: z.array(z.string()),
+        }),
+      ),
+    })
+    .superRefine((value, ctx) => {
+      console.log(value);
+      console.log(ctx);
+    }),
 });
 
-export type GardenCreateFormData = z.infer<typeof formSchema>;
+export const modalFormSchema = z.object({
+  fileContents: z.string().min(1, { message: "A file is required" }),
+
+  modal_functions: z
+    .array(
+      z.object({
+        title: z.string().min(1, { message: "Function title is required" }),
+        description: z.string(),
+        function_name: z.string().min(1, { message: "Function name is required" }),
+        year: z.string().min(4, { message: "Year must be 4 digits" }),
+        doi: z.string().min(1, {
+          message: "DOI is required",
+        }),
+        function_text: z.string().min(1, { message: "Function text is required" }),
+        authors: z.array(z.string()),
+        tags: z.array(z.string()),
+        test_functions: z.array(z.string()),
+      }),
+    )
+    .min(1, { message: "At least one function is required." }),
+  title: z.string().min(1, { message: "Garden Title is required" }),
+  app_name: z.string().min(1, { message: "App name is required" }),
+  description: z.string().min(1, { message: "Garden Description is required" }),
+});
+
+export type ModalUploadFormData = z.infer<typeof modalFormSchema>;
+export type GardenCreateFormData = z.infer<typeof gardenFormSchema>;
