--- garden/src/components/form/modal/create/ModalUploadPage.tsx
@@ -1,9 +1,0 @@
-import { ModalUploadForm } from "./ModalUploadForm";
-
-export const ModalUploadPage = () => {
-  return (
-    <div className="mx-auto max-w-4xl py-12 font-display">
-      <ModalUploadForm />
-    </div>
-  );
-};
